requirements.txt
requests
python-magic
grpcio
grpcio-tools
sentry-sdk
setuptools
PyJWT